+++ test/jest/security.test.tsx
@@ -16,13 +16,18 @@ import { act } from 'react-dom/test-utils';
 import { MemoryRouter } from 'react-router-dom';
 import Security from '../../src/Security';
 import { useOktaAuth } from '../../src/OktaContext';
+import { AuthState, OktaAuth } from '@okta/okta-auth-js';
+
+declare global {
+  let SKIP_VERSION_CHECK: any;
+}
 
 console.warn = jest.fn();
 
 describe('<Security />', () => {
-  let oktaAuth;
-  let initialAuthState;
-  const restoreOriginalUri = async (_, url) => {
+  let oktaAuth: OktaAuth;
+  let initialAuthState: AuthState | null;
+  const restoreOriginalUri = async (_: OktaAuth, url: string) => {
     location.href = url;
   };
   beforeEach(() => {
@@ -36,16 +41,18 @@ describe('<Security />', () => {
         addEnvironment: jest.fn(),
         getHttpHeader: jest.fn(),
         getVersion: jest.fn()
-      },
+      } as any,
       options: {},
       authStateManager: {
         getAuthState: jest.fn().mockImplementation(() => initialAuthState),
+        updateAuthState: jest.fn(),
         subscribe: jest.fn(),
         unsubscribe: jest.fn(),
-      },
+      } as any,
       start: jest.fn(),
       stop: jest.fn(),
-    };
+      isLoginRedirect: jest.fn().mockImplementation(() => false),
+    } as any;
   });
 
   it('adds an environmemnt to oktaAuth\'s _oktaUserAgent', () => {
@@ -61,7 +68,7 @@ describe('<Security />', () => {
   });
 
   it('logs a warning in case _oktaUserAgent is not available on auth SDK instance', () => {
-    const oktaAuthWithoutUserAgent = {
+    const oktaAuthWithoutUserAgent: any = {
       ...oktaAuth
     };
     delete oktaAuthWithoutUserAgent['_oktaUserAgent'];
@@ -74,7 +81,7 @@ describe('<Security />', () => {
   });
 
   describe('throws version not match error', () => {
-    let originalConsole;
+    let originalConsole: any;
 
     // turn off SKIP_VERSION_CHECK to test the functionality
     beforeEach(() => {
@@ -98,8 +105,8 @@ describe('<Security />', () => {
         _oktaUserAgent: {
           addEnvironment: jest.fn(),
           getVersion: jest.fn().mockReturnValue('1.0.0') // intentional large mock version
-        }
-      };
+        } as any
+      } as OktaAuth;
 
       const mockProps = {
         oktaAuth: oktaAuthWithMismatchingSDKVersion,
@@ -120,7 +127,7 @@ describe('<Security />', () => {
       };
 
       const mockProps = {
-        oktaAuth: oktaAuthWithMismatchingSDKVersion,
+        oktaAuth: oktaAuthWithMismatchingSDKVersion as any,
         restoreOriginalUri
       };
 
@@ -160,15 +167,16 @@ describe('<Security />', () => {
     expect(MyComponent).toHaveBeenCalled();
   });
 
-  it('calls start and updates the context', () => {
+  it('calls start and updates the context, does NOT call stop', () => {
+    initialAuthState = null;
     const newAuthState = {
       fromUpdateAuthState: true
     };
-    let callback;
-    oktaAuth.authStateManager.subscribe.mockImplementation(fn => {
+    let callback: (state: AuthState | null) => void;
+    oktaAuth.authStateManager.subscribe = jest.fn().mockImplementation(fn => {
       callback = fn;
     });
-    oktaAuth.start.mockImplementation(() => {
+    oktaAuth.start = jest.fn().mockImplementation(() => {
       callback(newAuthState);
     });
     const mockProps = {
@@ -190,7 +198,7 @@ describe('<Security />', () => {
         return null;
       });
 
-    mount(
+    const component = mount(
       <MemoryRouter>
         <Security {...mockProps}>
           <MyComponent />
@@ -201,10 +209,13 @@ describe('<Security />', () => {
     expect(oktaAuth.authStateManager.subscribe).toHaveBeenCalledTimes(1);
     expect(oktaAuth.start).toHaveBeenCalledTimes(1);
     expect(MyComponent).toHaveBeenCalledTimes(2);
+    component.unmount();
+    expect(oktaAuth.authStateManager.unsubscribe).toHaveBeenCalledTimes(1);
+    expect(oktaAuth.stop).toHaveBeenCalledTimes(0);
   });
 
   it('subscribes to "authStateChange" and updates the context', () => {
-    const mockAuthStates = [
+    const mockAuthStates: Array<AuthState | null> = [
       initialAuthState,
       {
         fromUpdateAuthState: true
@@ -213,24 +224,24 @@ describe('<Security />', () => {
         fromEventDispatch: true
       }
     ];
-    const callbacks = [];
+    const callbacks: Array<(state: AuthState | null) => void> = [];
     let stateCount = 0;
     callbacks.push(() => {
       // dummy subscriber that should be preserved after `<Security />` unmount
     });
-    oktaAuth.authStateManager.getAuthState.mockImplementation( () => { 
+    oktaAuth.authStateManager.getAuthState = jest.fn().mockImplementation( () => { 
       return mockAuthStates[stateCount];
     });
-    oktaAuth.authStateManager.subscribe.mockImplementation(fn => {
+    oktaAuth.authStateManager.subscribe = jest.fn().mockImplementation(fn => {
       callbacks.push(fn);
     });
-    oktaAuth.authStateManager.unsubscribe.mockImplementation(fn => {
+    oktaAuth.authStateManager.unsubscribe = jest.fn().mockImplementation(fn => {
       const index = callbacks.indexOf(fn);
       if (index !== -1) {
         callbacks.splice(index, 1);
       }
     });
-    oktaAuth.start.mockImplementation(() => {
+    oktaAuth.start = jest.fn().mockImplementation(() => {
       stateCount++;
       callbacks.map(fn => fn(mockAuthStates[stateCount]));
     });
@@ -265,6 +276,7 @@ describe('<Security />', () => {
         </Security>
       </MemoryRouter>
     );
+
     expect(callbacks.length).toEqual(2);
     expect(oktaAuth.authStateManager.subscribe).toHaveBeenCalledTimes(1);
     expect(oktaAuth.start).toHaveBeenCalledTimes(1);
@@ -277,7 +289,8 @@ describe('<Security />', () => {
     expect(MyComponent).toHaveBeenCalledTimes(1);
 
     component.unmount();
-    expect(oktaAuth.stop).toHaveBeenCalledTimes(1);
+    expect(oktaAuth.stop).toHaveBeenCalledTimes(0);
+    expect(oktaAuth.authStateManager.unsubscribe).toHaveBeenCalledTimes(1);
     expect(callbacks.length).toEqual(1);
   });
 
@@ -357,7 +370,7 @@ describe('<Security />', () => {
 
     it('should render error if oktaAuth props is not provided', () => {
       const mockProps = {
-        oktaAuth: null,
+        oktaAuth: null as any,
         restoreOriginalUri
       };
       const wrapper = mount(
@@ -371,7 +384,7 @@ describe('<Security />', () => {
     it('should render error if restoreOriginalUri prop is not provided', () => {
       const mockProps = {
         oktaAuth,
-        restoreOriginalUri: null
+        restoreOriginalUri: null as any
       };
       const wrapper = mount(
         <Security {...mockProps}>